requirements.txt
psycopg2-binary>=2.8.4
Jinja2>=2.10.1
requests>=2.22.0
pyyaml>=5.3
iso3166>=1.0.1
